First attempt at building pylons *before* getting supply blocked.
Causes a race condition if we have 100 minerals when the probe sets off to build the pylon, but then we dip under 100 en-route (by building another unit, for e.g.). Probe will attempt to build, fail, but will think it is done and go home.

Need some way to check the success of a build and keep trying until it is successful.